# openapi/shepherd.yaml
openapi: 3.1.0
info:
  title: shepherd API
  description: >-
    Health-only scaffold contract for the shepherd daemon. This is step 000
    of the stable-v1 plan: the spec is the product boundary — anything not
    in this contract does not exist. Later steps replace it with the full
    v1 surface; the authoritative contracts live in plan/contracts/.
  version: 0.1.0
  contact:
    name: apkg
    url: https://github.com/apkg-ai/shepherd
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:7437
    description: >-
      Local daemon (default port 7437, override with --port or
      SHEPHERD_PORT). v1 runs plain HTTP on localhost; TLS is a
      post-v1 concern (F4).
    x-internal: true
security:
  - bearerAuth: []
  - ownerSession: []
...
paths:
...
  /api/v1/agents/{agent_id}/revoke:
    post:
      operationId: revokeAgent
      tags:
        - identity
      summary: Revoke agent
      description: See plan workflow and command catalog for authoritative preconditions and effects.
      x-role: human
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            minLength: 36
            maxLength: 36
          description: agent id
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
            minLength: 36
            maxLength: 36
          description: Idempotency-Key
        - name: X-CSRF-Token
          in: header
          required: false
          description: Required for cookie-authenticated mutations; bearer clients omit.
          schema:
            type: string
            maxLength: 100
      responses:
        "200":
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ack"
              example:
                ok: false
        "400":
          description: Malformed request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Authentication required
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Resource not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: Workflow or idempotency conflict
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "413":
          description: Body too large
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "415":
          description: Unsupported media type
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "422":
          description: Validation failed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "428":
          description: Required command header missing
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Unexpected failure
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "503":
          description: Database unavailable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReasonInput"
            example:
              reason: Example
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Owner or agent bearer token: 32 random bytes, base64url, issued at
        bootstrap (owner-token file) or by createAgent. Opaque, not JWT.
        See plan/12-security-and-local-identity.md.
    ownerSession:
      type: apiKey
      in: cookie
      name: shepherd_session
      description: >-
        Owner browser session cookie issued by createBrowserSession.
        Cookie-authenticated mutations also require X-CSRF-Token.

  # ── Reusable headers ─────────────────────────────────────────────────
...
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      properties:
        type:
          type: string
          maxLength: 200
          description: >-
            Type. Exact semantics are specified in plan/03-domain-model.md and
            plan/04-workflow-state-machines.md.
        title:
          type: string
          maxLength: 200
          description: >-
            Title. Exact semantics are specified in plan/03-domain-model.md and
            plan/04-workflow-state-machines.md.
        status:
          type: integer
          minimum: 400
          maximum: 599
          description: >-
            Status. Exact semantics are specified in plan/03-domain-model.md and
            plan/04-workflow-state-machines.md.
          format: int32
        code:
          type: string
          maxLength: 80
          description: >-
            Code. Exact semantics are specified in plan/03-domain-model.md and
            plan/04-workflow-state-machines.md.
        detail:
          type: string
          maxLength: 2000
          description: >-
            Detail. Exact semantics are specified in plan/03-domain-model.md and
            plan/04-workflow-state-machines.md.
        request_id:
          type: string
          format: uuid
          description: >-
            Request id. Exact semantics are specified in plan/03-domain-model.md and
            plan/04-workflow-state-machines.md.
          minLength: 36
          maxLength: 36
        errors:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              field:
                type: string
                maxLength: 200
                description: >-
                  Field. Exact semantics are specified in plan/03-domain-model.md and
                  plan/04-workflow-state-machines.md.
              message:
                type: string
                maxLength: 500
                description: >-
                  Message. Exact semantics are specified in plan/03-domain-model.md and
                  plan/04-workflow-state-machines.md.
            required:
              - field
              - message
          maxItems: 16
          description: >-
            Errors. Exact semantics are specified in plan/03-domain-model.md and
            plan/04-workflow-state-machines.md.
          minItems: 0
      required:
        - type
        - title
        - status
        - code
        - detail
        - request_id
      description: Problem contract; see the domain and workflow specifications for invariants.
...
    Ack:
      type: object
      additionalProperties: false
      properties:
        ok:
          type: boolean
          description: >-
            Ok. Exact semantics are specified in plan/03-domain-model.md and
            plan/04-workflow-state-machines.md.
      required:
        - ok
      description: Ack contract; see the domain and workflow specifications for invariants.
    ReasonInput:
      type: object
      additionalProperties: false
      properties:
        reason:
          type: string
          maxLength: 2000
          minLength: 1
          description: >-
            Reason. Exact semantics are specified in plan/03-domain-model.md and
            plan/04-workflow-state-machines.md.
      required:
        - reason
      description: ReasonInput contract; see the domain and workflow specifications for invariants.